fix: move tomlconfig to config path & masterDB logic to SP
