--- cmd/migrate/toml/toml_test.go
@@ -1,1 +1,0 @@
-package tomlconfig
